Split the packages configuration data
Use the PackagesSelectionData structure for the software selection and
the PackagesConfigurationData structure for the DNF and RPM configuration.
