+++ tests/nosetests/pyanaconda_tests/modules/payloads/payload/module_payload_dnf_installation_test.py
@@ -23,7 +23,7 @@ from unittest.mock import patch, call
 
 from pyanaconda.core.constants import RPM_LANGUAGES_NONE
 from pyanaconda.core.util import join_paths
-from pyanaconda.modules.common.structures.payload import PackagesConfigurationData
+from pyanaconda.modules.common.structures.packages import PackagesConfigurationData
 from pyanaconda.modules.payloads.payload.dnf.installation import ImportRPMKeysTask, \
     SetRPMMacrosTask
